requirements.txt
gradio==4.31.5
langchain==0.2.1
langchain_community==0.2.1
python-dotenv==1.0.1
sentence-transformers==3.0.0
langchain_huggingface==0.2.1.